refactor angelfish tests for jira cloud numbering
