+++ tests/bdd/scale/test_NAS_T1287.py
@@ -17,9 +17,9 @@ import pytest
 pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1535.feature', 'Apps Page - Validate machinaris')
-def test_apps_page__validate_machinaris():
-    """Apps Page - Validate machinaris."""
+@scenario('features/NAS-T1287.feature', 'Apps Page - Validate Collabora')
+def test_apps_page__validate_collabora():
+    """Apps Page - Validate Collabora."""
 
 
 @given('the browser is open, navigate to the SCALE URL, and login')
@@ -49,28 +49,9 @@ def on_the_dashboard_click_on_apps(driver):
     driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
 
 
-@then('Stop Chia from running')
-def stop_chia_from_running(driver):
-    """Stop Chia from running."""
-    if is_element_present(driver, '//mat-ink-bar[@style="visibility: visible; left: 0px; width: 183px;"]') is False:
-        assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
-        driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-        assert wait_on_element(driver, 7, '//h3[contains(.,"No Applications Installed")]')
-    assert wait_on_element(driver, 20, '//mat-card[contains(.,"chia")]//span[contains(.,"Stop")]', 'clickable')
-    driver.find_element_by_xpath('//mat-card[contains(.,"chia")]//span[contains(.,"Stop")]').click()
-
-
-@then('Verify the application has stopped')
-def verify_the_application_has_stopped(driver):
-    """Verify the application has stopped."""
-    assert wait_on_element(driver, 5, '//h1[contains(.,"Stopping")]')
-    assert wait_on_element_disappear(driver, 60, '//h1[contains(.,"Stopping")]')
-    assert wait_on_element(driver, 15, '//mat-card[contains(.,"chia-test")]//span[contains(.,"STOPPED ")]')
-
-
-@then('open available applications')
-def open_available_applications(driver):
-    """open available applications."""
+@then('the Apps page load, open available applications')
+def the_apps_page_load_open_available_applications(driver):
+    """the Apps page load, open available applications."""
     assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
     driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
     assert wait_on_element(driver, 7, '//div[contains(.,"Available Applications")]')
@@ -80,35 +61,50 @@ def open_available_applications(driver):
 def click_install(driver):
     """click install."""
     time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    assert wait_on_element(driver, 20, '//mat-card[contains(.,"machinaris")]//span[contains(.,"Install")]', 'clickable')
-    driver.find_element_by_xpath('//mat-card[contains(.,"machinaris")]//span[contains(.,"Install")]').click()
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"collabora")]//span[contains(.,"Install")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"collabora")]//span[contains(.,"Install")]').click()
     assert wait_on_element(driver, 5, '//*[contains(.,"Please wait")]')
-    assert wait_on_element_disappear(driver, 30, '//*[contains(.,"Please wait")]')
+    assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
 
 
 @then('set application name')
 def set_application_name(driver):
     """set application name."""
-    assert wait_on_element(driver, 7, '//h3[contains(.,"machinaris")]')
+    assert wait_on_element(driver, 7, '//h3[contains(.,"collabora")]')
     assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
     driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
-    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('machinaris-test')
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('collabora-test')
     assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
     driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
 
 
+@then('set collabora configuration')
+def set_collabora_configuration(driver):
+    """set collabora configuration."""
+    driver.find_element_by_xpath('//input[@ix-auto="input__Password for WebUI"]').clear()
+    driver.find_element_by_xpath('//input[@ix-auto="input__Password for WebUI"]').send_keys('testingpass')
+
+    assert wait_on_element(driver, 5, '//mat-select[@ix-auto="select__Certificate"]', 'clickable')
+    driver.find_element_by_xpath('//mat-select[@ix-auto="select__Certificate"]').click()
+    # made the next xpath backward compatible
+    assert wait_on_element(driver, 7, '//span[contains(.,"freenas_default") or contains(.,"truenas_default")]', 'clickable')
+    driver.find_element_by_xpath('//span[contains(.,"freenas_default") or contains(.,"truenas_default")]').click()
+
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Collabora Configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Collabora Configuration"]').click()
+
+
+@then('set collabora environment variables')
+def set_collabora_environment_variables(driver):
+    """set collabora environment variables."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Collabora Environment Variables"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Collabora Environment Variables"]').click()
+
+
 @then('set networking')
 def set_networking(driver):
-    """set networking."""
     assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Networking"]', 'clickable')
     driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Networking"]').click()
-
-
-@then('set machinaris configuration')
-def set_machinaris_configuration(driver):
-    """set machinaris configuration."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Machinaris Configuration"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Machinaris Configuration"]').click()
 
 
 @then('set storage')
@@ -118,32 +114,8 @@ def set_storage(driver):
     driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
 
 
-@then('set Machinaris Environment Variables')
-def set_machinaris_environment_variables(driver):
-    """set Machinaris Environment Variables."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Machinaris Environment Variables"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Machinaris Environment Variables"]').click()
-
-
-@then('set Resource Limits')
-def set_resource_limits(driver):
-    """set Resource Limits."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Resource Limits"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Resource Limits"]').click()
-
-
-@then('set Configure Coins')
-def set_configure_coins(driver):
-    """set Configure Coins."""
-    assert wait_on_element(driver, 10, '//mat-checkbox[@ix-auto="checkbox__Enable Flax"]', 'clickable')
-    driver.find_element_by_xpath('//mat-checkbox[@ix-auto="checkbox__Enable Flax"]').click()
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Configure Coins"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Configure Coins"]').click()
-
-
 @then('confirm options')
 def confirm_options(driver):
-    """confirm options."""
     assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
     driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
 
@@ -157,16 +129,16 @@ def confirm_installation_is_successful(driver):
     assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
     driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
     time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    if is_element_present(driver, '//mat-card[contains(.,"machinaris-test")]//span[@class="status active"]') is False:
-        assert wait_on_element(driver, 20, '//strong[contains(.,"machinaris-test")]')
-        assert wait_on_element(driver, 20, '//strong[contains(.,"machinaris-test")]', 'clickable')
-        driver.find_element_by_xpath('//strong[contains(.,"machinaris-test")]').click()
+    if is_element_present(driver, '//mat-card[contains(.,"collabora-test")]//span[@class="status active"]') is False:
+        assert wait_on_element(driver, 20, '//strong[contains(.,"collabora-test")]')
+        assert wait_on_element(driver, 20, '//strong[contains(.,"collabora-test")]', 'clickable')
+        driver.find_element_by_xpath('//strong[contains(.,"collabora-test")]').click()
         if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
             assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
-        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"machinaris-test")]')
+        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"collabora-test")]')
         assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
         driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
-        while is_element_present(driver, '//div[(normalize-space(text())="Created container machinaris")]') is False:
+        while is_element_present(driver, '//div[(normalize-space(text())="Started container collabora")]') is False:
             time.sleep(2)
             assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
             driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
@@ -176,11 +148,12 @@ def confirm_installation_is_successful(driver):
         else:
             assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
             driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
-            time.sleep(45)  # Because of slow start up times, this takes another 10-20 second to switch from "Deploying to Active"  So we can either flip the page constantly or just wait and give it time.
+            time.sleep(1)  # wait for popup to close
+            # we have to change tab for UI to refresh
             assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
             driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
             assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
             driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-            assert wait_on_element(driver, 600, '//mat-card[contains(.,"machinaris-test")]//span[@class="status active"]')
+            assert wait_on_element(driver, 500, '//mat-card[contains(.,"collabora-test")]//span[@class="status active"]')
     else:
-        assert wait_on_element(driver, 600, '//mat-card[contains(.,"machinaris-test")]//span[@class="status active"]')
+        assert wait_on_element(driver, 500, '//mat-card[contains(.,"collabora-test")]//span[@class="status active"]')
