+++ tests/bdd/scale/test_NAS_T1334.py
@@ -1,0 +1,72 @@
+# coding=utf-8
+"""SCALE UI: feature tests."""
+
+from function import (
+    wait_on_element,
+    is_element_present,
+    wait_on_element_disappear,
+)
+from pytest_bdd import (
+    given,
+    scenario,
+    then,
+    when,
+)
+import pytest
+pytestmark = [pytest.mark.debug_test]
+
+
+@scenario('features/NAS-T1334.feature', 'Apps Page - Stop an App')
+def test_apps_page__stop_an_app():
+    """Apps Page - Stop an App."""
+
+
+@given('the browser is open, navigate to the SCALE URL, and login')
+def the_browser_is_open_navigate_to_the_scale_url_and_login(driver, nas_ip, root_password):
+    """the browser is open, navigate to the SCALE URL, and login."""
+    if nas_ip not in driver.current_url:
+        driver.get(f"http://{nas_ip}")
+        assert wait_on_element(driver, 10, '//input[@data-placeholder="Username"]')
+    if not is_element_present(driver, '//mat-list-item[@ix-auto="option__Dashboard"]'):
+        assert wait_on_element(driver, 10, '//input[@data-placeholder="Username"]')
+        driver.find_element_by_xpath('//input[@data-placeholder="Username"]').clear()
+        driver.find_element_by_xpath('//input[@data-placeholder="Username"]').send_keys('root')
+        driver.find_element_by_xpath('//input[@data-placeholder="Password"]').clear()
+        driver.find_element_by_xpath('//input[@data-placeholder="Password"]').send_keys(root_password)
+        assert wait_on_element(driver, 5, '//button[@name="signin_button"]')
+        driver.find_element_by_xpath('//button[@name="signin_button"]').click()
+    else:
+        assert wait_on_element(driver, 5, '//mat-list-item[@ix-auto="option__Dashboard"]', 'clickable')
+        driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Dashboard"]').click()
+
+
+@when('on the Dashboard, click on apps')
+def on_the_dashboard_click_on_apps(driver):
+    """on the Dashboard, click on apps."""
+    assert wait_on_element(driver, 10, '//span[contains(.,"Dashboard")]')
+    assert wait_on_element(driver, 10, '//mat-list-item[@ix-auto="option__Apps"]', 'clickable')
+    driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
+
+
+@then('the Apps page load, open installed applications')
+def the_apps_page_load_open_installed_applications(driver):
+    """the Apps page load, open installed applications."""
+    if is_element_present(driver, '//mat-ink-bar[@style="visibility: visible; left: 0px; width: 183px;"]') is False:
+        assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+        driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+        assert wait_on_element(driver, 7, '//h3[contains(.,"No Applications Installed")]')
+
+
+@then('click the stop button and confirm')
+def click_the_stop_button_and_confirm(driver):
+    """click the stop button and confirm."""
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"collabora")]//span[contains(.,"Stop")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"collabora")]//span[contains(.,"Stop")]').click()
+
+
+@then('Verify the application has stopped')
+def verify_the_application_has_stopped(driver):
+    """Verify the application has stopped."""
+    assert wait_on_element(driver, 5, '//h1[contains(.,"Stopping")]')
+    assert wait_on_element_disappear(driver, 60, '//h1[contains(.,"Stopping")]')
+    assert wait_on_element(driver, 15, '//mat-card[contains(.,"collabora-test")]//span[contains(.,"STOPPED ")]')
