+++ tests/bdd/scale/test_NAS_T1341.py
@@ -1,22 +1,25 @@
 # coding=utf-8
 """SCALE UI: feature tests."""
 
+import time
 from function import (
     wait_on_element,
     is_element_present,
-    wait_on_element_disappear
+    wait_on_element_disappear,
 )
 from pytest_bdd import (
     given,
     scenario,
     then,
-    when,
+    when
 )
+import pytest
+pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1341.feature', 'Verify a certificate authority can be deleted')
-def test_verify_a_certificate_authority_can_be_deleted():
-    """Verify a certificate authority can be deleted."""
+@scenario('features/NAS-T1341.feature', 'Apps Page - Validate ipfs')
+def test_apps_page__validate_ipfs():
+    """Apps Page - Validate ipfs."""
 
 
 @given('the browser is open, navigate to the SCALE URL, and login')
@@ -34,39 +37,117 @@ def the_browser_is_open_navigate_to_the_scale_url_and_login(driver, nas_ip, root
         assert wait_on_element(driver, 5, '//button[@name="signin_button"]')
         driver.find_element_by_xpath('//button[@name="signin_button"]').click()
     else:
-        assert wait_on_element(driver, 10, '//mat-list-item[@ix-auto="option__Dashboard"]', 'clickable')
+        assert wait_on_element(driver, 5, '//mat-list-item[@ix-auto="option__Dashboard"]', 'clickable')
         driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Dashboard"]').click()
 
 
-@when('on the Dashboard, click on credentials and certificates')
-def on_the_dashboard_click_on_credentials_and_certificates(driver):
-    """on the Dashboard, click on credentials and certificates."""
+@when('on the Dashboard, click on apps')
+def on_the_dashboard_click_on_apps(driver):
+    """on the Dashboard, click on apps."""
     assert wait_on_element(driver, 10, '//span[contains(.,"Dashboard")]')
-    assert wait_on_element(driver, 7, '//mat-list-item[@ix-auto="option__Credentials"]', 'clickable')
-    driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Credentials"]').click()
-    assert wait_on_element(driver, 7, '//*[contains(@class,"lidein-nav-md")]//mat-list-item[@ix-auto="option__Certificates"]', 'clickable')
-    driver.find_element_by_xpath('//*[contains(@class,"lidein-nav-md")]//mat-list-item[@ix-auto="option__Certificates"]').click()
+    assert wait_on_element(driver, 10, '//mat-list-item[@ix-auto="option__Apps"]', 'clickable')
+    driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
 
 
-@then('click on the trash icon for ca1')
-def click_on_the_trash_icon_for_ca1(driver):
-    """click on the trash icon for ca1."""
-    assert wait_on_element(driver, 7, '//h3[contains(text(),"Certificate Authorities")]')
-    assert wait_on_element(driver, 5, '//tr[contains(.,"ca1")]//mat-icon[contains(text(),"delete")]', 'clickable')
-    driver.find_element_by_xpath('//tr[contains(.,"ca1")]//mat-icon[contains(text(),"delete")]').click()
+@then('the Apps page load, open available applications')
+def the_apps_page_load_open_available_applications(driver):
+    """the Apps page load, open available applications."""
+    assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
+    driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
+    assert wait_on_element(driver, 7, '//div[contains(.,"Available Applications")]')
 
 
-@then('click the confirm checkbox and click delete')
-def click_the_confirm_checkbox_and_click_delete(driver):
-    """click the confirm checkbox and click delete."""
-    assert wait_on_element(driver, 5, '//h1[contains(.,"Delete")]')
-    assert wait_on_element(driver, 10, '//mat-checkbox[@ix-auto="checkbox__CONFIRM"]', 'clickable')
-    driver.find_element_by_xpath('//mat-checkbox[@ix-auto="checkbox__CONFIRM"]').click()
-    assert wait_on_element(driver, 10, '//button[@ix-auto="button__DELETE"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__DELETE"]').click()
+@then('click install')
+def click_install(driver):
+    """click install."""
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"ipfs")]//span[contains(.,"Install")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"ipfs")]//span[contains(.,"Install")]').click()
+    if is_element_present(driver, '//*[contains(.,"Please wait")]'):
+        assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
 
 
-@then('verify that the CA was deleted')
-def verify_that_the_ca_was_deleted(driver):
-    """verify that the CA was deleted."""
-    assert wait_on_element_disappear(driver, 20, '//li[contains(.,"Name: ca1")]')
+@then('set application name')
+def set_application_name(driver):
+    """set application name."""
+    assert wait_on_element(driver, 7, '//h3[contains(.,"ipfs")]')
+    assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('ipfs-test')
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
+
+
+@then('set workload configuration')
+def set_workload_configuration(driver):
+    """set workload configuration."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Workload Configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Workload Configuration"]').click()
+
+
+@then('set storage')
+def set_storage(driver):
+    """set storage."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Storage"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
+
+
+@then('IPFS Configuration')
+def ipfs_configuration(driver):
+    """IPFS Configuration."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_IPFS Configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_IPFS Configuration"]').click()
+
+
+@then('Advanced DNS Settings')
+def advanced_dns_settings(driver):
+    """Advanced DNS Settings."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Advanced DNS Settings"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Advanced DNS Settings"]').click()
+
+
+@then('confirm options')
+def confirm_options(driver):
+    """confirm options."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
+
+    assert wait_on_element(driver, 5, '//*[contains(.,"Installing")]')
+    assert wait_on_element_disappear(driver, 45, '//*[contains(.,"Installing")]')
+
+
+@then('confirm installation is successful')
+def confirm_installation_is_successful(driver):
+    """confirm installation is successful."""
+    assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+    driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    if is_element_present(driver, '//mat-card[contains(.,"ipfs-test")]//span[@class="status active"]') is False:
+        assert wait_on_element(driver, 20, '//strong[contains(.,"ipfs-test")]')
+        assert wait_on_element(driver, 20, '//strong[contains(.,"ipfs-test")]', 'clickable')
+        driver.find_element_by_xpath('//strong[contains(.,"ipfs-test")]').click()
+        assert wait_on_element(driver, 5, '//*[contains(.,"Please wait")]')
+        if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
+            assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"ipfs-test")]')
+        assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
+        driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
+        while is_element_present(driver, '//div[(normalize-space(text())="Started container ipfs")]') is False:
+            time.sleep(2)
+            assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
+            # make sure Please wait pop up is gone before continuing.
+            if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
+                assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        else:
+            assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
+            time.sleep(1)  # wait for popup to close
+            # we have to change tab for UI to refresh
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+            assert wait_on_element(driver, 500, '//mat-card[contains(.,"ipfs-test")]//span[@class="status active"]')
+    else:
+        assert wait_on_element(driver, 500, '//mat-card[contains(.,"ipfs-test")]//span[@class="status active"]')
