+++ tests/bdd/scale/test_NAS_T1343.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 """SCALE UI: feature tests."""
 
+import time
 from function import (
     wait_on_element,
     is_element_present,
@@ -16,9 +17,9 @@ import pytest
 pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1343.feature', 'Apps Page Validation')
-def test_apps_page_validation():
-    """Apps Page Validation."""
+@scenario('features/NAS-T1343.feature', 'Apps Page - Validate nextcloud')
+def test__apps_page__validate_nextcloud():
+    """SCALE UI: Apps Page - Validate nextcloud."""
 
 
 @given('the browser is open, navigate to the SCALE URL, and login')
@@ -48,105 +49,112 @@ def on_the_dashboard_click_on_apps(driver):
     driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
 
 
-@then('the Apps page load, select pool')
-def the_apps_page_load_select_pool(driver):
-    """the Apps page load, select pool."""
+@then('the Apps page load, open available applications')
+def the_apps_page_load_open_available_applications(driver):
+    """the Apps page load, open available applications."""
+    assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
+    driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
     assert wait_on_element(driver, 7, '//div[contains(.,"Available Applications")]')
-    assert wait_on_element(driver, 7, '//h1[contains(.,"Choose a pool for Apps")]')
-    assert wait_on_element(driver, 5, '//mat-select[@ix-auto="select__Pools"]', 'clickable')
-    driver.find_element_by_xpath('//mat-select[@ix-auto="select__Pools"]').click()
-    assert wait_on_element(driver, 7, '//mat-option[@ix-auto="option__Pools_tank"]', 'clickable')
-    driver.find_element_by_xpath('//mat-option[@ix-auto="option__Pools_tank"]').click()
-    assert wait_on_element(driver, 7, '//button[@name="Choose_button"]', 'clickable')
-    driver.find_element_by_xpath('//button[@name="Choose_button"]').click()
-    assert wait_on_element_disappear(driver, 60, '//h1[contains(.,"Configuring...")]')
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__CLOSE"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__CLOSE"]').click()
 
 
-@then('the Available Applications Tab loads')
-def the_available_applications_tab_loads(driver):
-    """the Available Applications Tab loads."""
-    # used for local testing, so you dont have to unset and reset the pool every time 
-    # assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
-    # driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
-    assert wait_on_element(driver, 7, '//h3[contains(.,"minio")]')
+@then('click install')
+def click_install(driver):
+    """click install."""
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"nextcloud")]//span[contains(.,"Install")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"nextcloud")]//span[contains(.,"Install")]').click()
+    if is_element_present(driver, '//*[contains(.,"Please wait")]'):
+        assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
 
 
-@then('verify the setting slide out works')
-def verify_the_setting_slide_out_works(driver):
-    """verify the setting slide out works."""
-    assert wait_on_element(driver, 10, '//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]').click()
-    assert wait_on_element(driver, 10, '//span[contains(text(),"Advanced Settings")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(text(),"Advanced Settings")]').click()
-    assert wait_on_element(driver, 7, '//h3[contains(.,"Kubernetes Settings")]')
-    assert wait_on_element(driver, 10, '//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]', 'clickable')
-    driver.find_element_by_xpath('//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]').click()
+@then('set application name')
+def set_application_name(driver):
+    """set application name."""
+    assert wait_on_element(driver, 7, '//h3[contains(.,"nextcloud")]')
+    assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('nextcloud-test')
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
 
 
-@then('open the Installed Applications page')
-def open_the_installed_applications_page(driver):
-    """open the Installed Applications page."""
+@then('set nexcloud configuration')
+def set_nexcloud_configuration(driver):
+    """set nexcloud configuration."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Nextcloud Configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Nextcloud Configuration"]').click()
+
+
+@then('set storage')
+def set_storage(driver):
+    """set storage."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Storage"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
+
+
+@then('on CronJob configuration click next')
+def on_cronjob_configuration_click_next(driver):
+    """on CronJob configuration click next."""
+    wait_on_element(driver, 5, '//mat-step-header[contains(.,"CronJob configuration") and @tabindex="0"]')
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_CronJob configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_CronJob configuration"]').click()
+
+
+@then('set Scaling Upgrade Policy')
+def set_scaling_upgrade_policy(driver):
+    """set Scaling Upgrade Policy."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Scaling/Upgrade Policy"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Scaling/Upgrade Policy"]').click()
+
+
+@then('Advanced DNS Settings')
+def advanced_dns_settings(driver):
+    """Advanced DNS Settings."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Advanced DNS Settings"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Advanced DNS Settings"]').click()
+
+
+@then('confirm options')
+def confirm_options(driver):
+    """confirm options."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
+
+    assert wait_on_element(driver, 10, '//*[contains(.,"Installing")]')
+    assert wait_on_element_disappear(driver, 60, '//*[contains(.,"Installing")]')
+
+
+@then('confirm installation is successful')
+def confirm_installation_is_successful(driver):
+    """confirm installation is successful."""
     assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
     driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-    assert wait_on_element(driver, 7, '//h3[contains(.,"No Applications Installed")]')
-    # test the buttons
-    assert wait_on_element(driver, 10, '//span[contains(.,"Launch Docker Image")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(.,"Launch Docker Image")]').click()
-    if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
-        assert wait_on_element_disappear(driver, 120, '//*[contains(.,"Please wait")]')
-    assert wait_on_element(driver, 10, '//mat-icon[contains(.,"cancel")]', 'clickable')
-    driver.find_element_by_xpath('//mat-icon[contains(.,"cancel")]').click()
-    assert wait_on_element(driver, 10, '//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]').click()
-    assert wait_on_element(driver, 10, '//span[contains(text(),"Advanced Settings")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(text(),"Advanced Settings")]').click()
-    assert wait_on_element(driver, 7, '//h3[contains(.,"Kubernetes Settings")]')
-    assert wait_on_element(driver, 10, '//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]', 'clickable')
-    driver.find_element_by_xpath('//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]').click()
-
-
-@then('open the Manage Docker Images Page')
-def open_the_manage_docker_images_page(driver):
-    """open the Manage Docker Images Page."""
-    assert wait_on_element(driver, 10, '//div[contains(text(),"Manage Docker Images")]', 'clickable')
-    driver.find_element_by_xpath('//div[contains(text(),"Manage Docker Images")]').click()
-    # seems like sometimes zfs-driver is present.
-    assert wait_on_element(driver, 5, '//h3[contains(.,"No Docker Images")]') or wait_on_element(driver, 5, '//div[contains(.,"rancher")]')
-    # test the buttons
-    assert wait_on_element(driver, 10, '//span[contains(.,"Launch Docker Image")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(.,"Launch Docker Image")]').click()
-    if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
-        assert wait_on_element_disappear(driver, 120, '//*[contains(.,"Please wait")]')
-    assert wait_on_element(driver, 10, '//mat-icon[contains(.,"cancel")]', 'clickable')
-    driver.find_element_by_xpath('//mat-icon[contains(.,"cancel")]').click()
-    assert wait_on_element(driver, 10, '//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]').click()
-    assert wait_on_element(driver, 10, '//span[contains(text(),"Advanced Settings")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(text(),"Advanced Settings")]').click()
-    assert wait_on_element(driver, 7, '//h3[contains(.,"Kubernetes Settings")]')
-    assert wait_on_element(driver, 10, '//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]', 'clickable')
-    driver.find_element_by_xpath('//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]').click()
-
-
-@then('open the Manage Catalogs Page')
-def open_the_manage_catalogs_page(driver):
-    """open the Manage Catalogs Page."""
-    assert wait_on_element(driver, 10, '//div[contains(text(),"Manage Catalogs")]', 'clickable')
-    driver.find_element_by_xpath('//div[contains(text(),"Manage Catalogs")]').click()
-    assert wait_on_element(driver, 7, '//div[contains(.,"https://github.com/truenas/charts.git")]')
-    # test the buttons
-    assert wait_on_element(driver, 10, '//span[contains(.,"Launch Docker Image")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(.,"Launch Docker Image")]').click()
-    if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
-        assert wait_on_element_disappear(driver, 120, '//*[contains(.,"Please wait")]')
-    assert wait_on_element(driver, 10, '//mat-icon[contains(.,"cancel")]', 'clickable')
-    driver.find_element_by_xpath('//mat-icon[contains(.,"cancel")]').click()
-    assert wait_on_element(driver, 10, '//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]').click()
-    assert wait_on_element(driver, 10, '//span[contains(text(),"Advanced Settings")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(text(),"Advanced Settings")]').click()
-    assert wait_on_element(driver, 7, '//h3[contains(.,"Kubernetes Settings")]')
-    assert wait_on_element(driver, 10, '//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]', 'clickable')
-    driver.find_element_by_xpath('//div[@class="ix-slidein-title-bar"]//mat-icon[contains(.,"cancel")]').click()
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    if is_element_present(driver, '//mat-card[contains(.,"nextcloud-test")]//span[@class="status active"]') is False:
+        assert wait_on_element(driver, 20, '//strong[contains(.,"nextcloud-test")]')
+        assert wait_on_element(driver, 20, '//strong[contains(.,"nextcloud-test")]', 'clickable')
+        driver.find_element_by_xpath('//strong[contains(.,"nextcloud-test")]').click()
+        if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
+            assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"nextcloud-test")]')
+        assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
+        driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
+        while is_element_present(driver, '//div[(normalize-space(text())="Started container nextcloud")]') is False:
+            time.sleep(2)
+            assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
+            # make sure Please wait pop up is gone before continuing.
+            if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
+                assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        else:
+            assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
+            time.sleep(1)  # wait for popup to close
+            # we have to change tab for UI to refresh
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+            assert wait_on_element(driver, 500, '//mat-card[contains(.,"nextcloud-test")]//span[@class="status active"]')
+    else:
+        assert wait_on_element(driver, 500, '//mat-card[contains(.,"nextcloud-test")]//span[@class="status active"]')
