+++ tests/bdd/scale/test_NAS_T1344.py
@@ -5,21 +5,21 @@ import time
 from function import (
     wait_on_element,
     is_element_present,
-    wait_on_element_disappear
+    wait_on_element_disappear,
 )
 from pytest_bdd import (
     given,
     scenario,
     then,
-    when
+    when,
 )
 import pytest
 pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1344.feature', 'Apps Page remove and readd pool')
-def test_apps_page_remove_and_readd_pool():
-    """Apps Page remove and readd pool."""
+@scenario('features/NAS-T1344.feature', 'Apps Page - Validate chia')
+def test_apps_page__validate_chia():
+    """Apps Page - Validate chia."""
 
 
 @given('the browser is open, navigate to the SCALE URL, and login')
@@ -47,51 +47,106 @@ def on_the_dashboard_click_on_apps(driver):
     assert wait_on_element(driver, 10, '//span[contains(.,"Dashboard")]')
     assert wait_on_element(driver, 10, '//mat-list-item[@ix-auto="option__Apps"]', 'clickable')
     driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
-    assert wait_on_element(driver, 10, '//h1[text()="Applications"]')
+
+
+@then('the Apps page load, open available applications')
+def the_apps_page_load_open_available_applications(driver):
+    """the Apps page load, open available applications."""
     assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
     driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
-    # Wait for Available Applications UI to load
-    assert wait_on_element(driver, 60, '//h3[text()="plex"]')
-    assert wait_on_element(driver, 30, '//button[@ix-auto="button__plex-install"]', 'clickable')
-    # Sleep to make sure that the drop does not disappear
-    time.sleep(3)
-
-
-@then('the Apps page load, click settings, unset pool')
-def the_apps_page_load_click_settings_unset_pool(driver):
-    """the Apps page load, click settings, unset pool."""
-    assert wait_on_element(driver, 20, '//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]').click()
-    assert wait_on_element(driver, 5, '//span[contains(text(),"Unset Pool")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(text(),"Unset Pool")]').click()
-    assert wait_on_element(driver, 20, '//h1[contains(.,"Unset Pool")]')
-
-
-@then('confirm unset pool and wait')
-def confirm_unset_pool_and_wait(driver):
-    """confirm unset pool and wait."""
-    assert wait_on_element(driver, 10, '//span[contains(text(),"UNSET")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(text(),"UNSET")]').click()
-    assert wait_on_element_disappear(driver, 60, '//h1[contains(.,"Configuring...")]')
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__CLOSE"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__CLOSE"]').click()
-
-
-@then('click setting, reset pool')
-def click_setting_reset_pool(driver):
-    """click setting, reset pool."""
-    assert wait_on_element(driver, 10, '//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto-type="button"]//span[contains(text(),"Settings")]').click()
-    assert wait_on_element(driver, 10, '//span[contains(text(),"Choose Pool")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(text(),"Choose Pool")]').click()
-    assert wait_on_element(driver, 7, '//h1[contains(.,"Choose a pool for Apps")]')
-    assert wait_on_element(driver, 5, '//mat-select[@ix-auto="select__Pools"]', 'clickable')
-    driver.find_element_by_xpath('//mat-select[@ix-auto="select__Pools"]').click()
-    assert wait_on_element(driver, 7, '//mat-option[@ix-auto="option__Pools_tank"]', 'clickable')
-    driver.find_element_by_xpath('//mat-option[@ix-auto="option__Pools_tank"]').click()
-    assert wait_on_element(driver, 7, '//button[@name="Choose_button"]', 'clickable')
-    driver.find_element_by_xpath('//button[@name="Choose_button"]').click()
-    assert wait_on_element_disappear(driver, 60, '//h1[contains(.,"Configuring...")]')
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__CLOSE"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__CLOSE"]').click()
     assert wait_on_element(driver, 7, '//div[contains(.,"Available Applications")]')
+
+
+@then('click install')
+def click_install(driver):
+    """click install."""
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"chia")]//span[contains(.,"Install")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"chia")]//span[contains(.,"Install")]').click()
+    if is_element_present(driver, '//*[contains(.,"Please wait")]'):
+        assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+
+
+@then('set application name')
+def set_application_name(driver):
+    """set application name."""
+    assert wait_on_element(driver, 7, '//h3[contains(.,"chia")]')
+    assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('chia-test')
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
+
+
+@then('set storage')
+def set_storage(driver):
+    """set storage."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Storage"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
+
+
+@then('set farmr.net configuration')
+def set_farmrnet_configuration(driver):
+    """set farmr.net configuration."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Farmr.net Configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Farmr.net Configuration"]').click()
+
+
+@then('Chia Environment Variables')
+def chia_environment_variables(driver):
+    """Chia Environment Variables."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Chia Environment Variables"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Chia Environment Variables"]').click()
+
+
+@then('networking')
+def networking(driver):
+    """networking."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Networking"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Networking"]').click()
+
+
+@then('confirm options')
+def confirm_options(driver):
+    """confirm options."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
+
+    assert wait_on_element(driver, 5, '//*[contains(.,"Installing")]')
+    assert wait_on_element_disappear(driver, 120, '//*[contains(.,"Installing")]')
+
+
+@then('confirm installation is successful')
+def confirm_installation_is_successful(driver):
+    """confirm installation is successful."""
+    assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+    driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    if is_element_present(driver, '//mat-card[contains(.,"chia-test")]//span[@class="status active"]') is False:
+        assert wait_on_element(driver, 20, '//strong[contains(.,"chia-test")]')
+        assert wait_on_element(driver, 20, '//strong[contains(.,"chia-test")]', 'clickable')
+        driver.find_element_by_xpath('//strong[contains(.,"chia-test")]').click()
+        if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
+            assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"chia-test")]')
+        assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
+        driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
+        while is_element_present(driver, '//div[(normalize-space(text())="Started container chia")]') is False:
+            time.sleep(2)
+            assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
+            # make sure Please wait pop up is gone before continuing.
+            if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
+                assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        else:
+            assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
+            time.sleep(1)  # wait for popup to close
+            # we have to change tab for UI to refresh
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+            assert wait_on_element(driver, 500, '//mat-card[contains(.,"chia-test")]//span[@class="status active"]')
+    else:
+        assert wait_on_element(driver, 500, '//mat-card[contains(.,"chia-test")]//span[@class="status active"]')
