+++ tests/bdd/scale/test_NAS_T1345.py
@@ -11,15 +11,15 @@ from pytest_bdd import (
     given,
     scenario,
     then,
-    when
+    when,
 )
 import pytest
 pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1345.feature', 'Apps Page - Validate Collabora')
-def test_apps_page__validate_collabora():
-    """Apps Page - Validate Collabora."""
+@scenario('features/NAS-T1345.feature', 'Apps Page - Validate plex')
+def test_apps_page__validate_plex():
+    """Apps Page - Validate plex."""
 
 
 @given('the browser is open, navigate to the SCALE URL, and login')
@@ -61,48 +61,33 @@ def the_apps_page_load_open_available_applications(driver):
 def click_install(driver):
     """click install."""
     time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    assert wait_on_element(driver, 20, '//mat-card[contains(.,"collabora")]//span[contains(.,"Install")]', 'clickable')
-    driver.find_element_by_xpath('//mat-card[contains(.,"collabora")]//span[contains(.,"Install")]').click()
-    assert wait_on_element(driver, 5, '//*[contains(.,"Please wait")]')
-    assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"plex")]//span[contains(.,"Install")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"plex")]//span[contains(.,"Install")]').click()
+    if is_element_present(driver, '//*[contains(.,"Please wait")]'):
+        assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
 
 
 @then('set application name')
 def set_application_name(driver):
     """set application name."""
-    assert wait_on_element(driver, 7, '//h3[contains(.,"collabora")]')
+    assert wait_on_element(driver, 7, '//h3[contains(.,"plex")]')
     assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
     driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
-    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('collabora-test')
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('plex-test')
     assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
     driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
 
 
-@then('set collabora configuration')
-def set_collabora_configuration(driver):
-    """set collabora configuration."""
-    driver.find_element_by_xpath('//input[@ix-auto="input__Password for WebUI"]').clear()
-    driver.find_element_by_xpath('//input[@ix-auto="input__Password for WebUI"]').send_keys('testingpass')
-
-    assert wait_on_element(driver, 5, '//mat-select[@ix-auto="select__Certificate"]', 'clickable')
-    driver.find_element_by_xpath('//mat-select[@ix-auto="select__Certificate"]').click()
-    # made the next xpath backward compatible
-    assert wait_on_element(driver, 7, '//span[contains(.,"freenas_default") or contains(.,"truenas_default")]', 'clickable')
-    driver.find_element_by_xpath('//span[contains(.,"freenas_default") or contains(.,"truenas_default")]').click()
-
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Collabora Configuration"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Collabora Configuration"]').click()
-
-
-@then('set collabora environment variables')
-def set_collabora_environment_variables(driver):
-    """set collabora environment variables."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Collabora Environment Variables"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Collabora Environment Variables"]').click()
+@then('set plex configuration')
+def set_plex_configuration(driver):
+    """set plex configuration."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Plex Configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Plex Configuration"]').click()
 
 
 @then('set networking')
 def set_networking(driver):
+    """set networking."""
     assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Networking"]', 'clickable')
     driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Networking"]').click()
 
@@ -114,8 +99,30 @@ def set_storage(driver):
     driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
 
 
+@then('set Scaling/Upgrade Policy')
+def set_scalingupgrade_policy(driver):
+    """set Scaling/Upgrade Policy."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Scaling/Upgrade Policy"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Scaling/Upgrade Policy"]').click()
+
+
+@then('set Resource Reservation')
+def set_resource_reservation(driver):
+    """set Resource Reservation."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Resource Reservation"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Resource Reservation"]').click()
+
+
+@then('set Advanced DNS Settings')
+def set_advanced_dns_settings(driver):
+    """set Advanced DNS Settings."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Advanced DNS Settings"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Advanced DNS Settings"]').click()
+
+
 @then('confirm options')
 def confirm_options(driver):
+    """confirm options."""
     assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
     driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
 
@@ -126,19 +133,20 @@ def confirm_options(driver):
 @then('confirm installation is successful')
 def confirm_installation_is_successful(driver):
     """confirm installation is successful."""
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
     assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
     driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    if is_element_present(driver, '//mat-card[contains(.,"collabora-test")]//span[@class="status active"]') is False:
-        assert wait_on_element(driver, 20, '//strong[contains(.,"collabora-test")]')
-        assert wait_on_element(driver, 20, '//strong[contains(.,"collabora-test")]', 'clickable')
-        driver.find_element_by_xpath('//strong[contains(.,"collabora-test")]').click()
+    time.sleep(2)  # we have to wait for the page to settle down and the cards to fully load
+    if is_element_present(driver, '//mat-card[contains(.,"plex-test")]//span[@class="status active"]') is False:
+        assert wait_on_element(driver, 20, '//strong[contains(.,"plex-test")]')
+        assert wait_on_element(driver, 20, '//strong[contains(.,"plex-test")]', 'clickable')
+        driver.find_element_by_xpath('//strong[contains(.,"plex-test")]').click()
         if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
             assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
-        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"collabora-test")]')
+        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"plex-test")]')
         assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
         driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
-        while is_element_present(driver, '//div[(normalize-space(text())="Started container collabora")]') is False:
+        while is_element_present(driver, '//div[(normalize-space(text())="Started container plex")]') is False:
             time.sleep(2)
             assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
             driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
@@ -148,12 +156,11 @@ def confirm_installation_is_successful(driver):
         else:
             assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
             driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
-            time.sleep(1)  # wait for popup to close
-            # we have to change tab for UI to refresh
+            time.sleep(30)  # Because of slow start up times, Plex takes another 10-15 second to switch from "Deploying to Active"  So we can either flip the page constantly or just wait and give it time.
             assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
             driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
             assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
             driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-            assert wait_on_element(driver, 500, '//mat-card[contains(.,"collabora-test")]//span[@class="status active"]')
+            assert wait_on_element(driver, 500, '//mat-card[contains(.,"plex-test")]//span[@class="status active"]')
     else:
-        assert wait_on_element(driver, 500, '//mat-card[contains(.,"collabora-test")]//span[@class="status active"]')
+        assert wait_on_element(driver, 500, '//mat-card[contains(.,"plex-test")]//span[@class="status active"]')
