+++ tests/bdd/scale/test_NAS_T1346.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 """SCALE UI: feature tests."""
 
+import time
 from function import (
     wait_on_element,
     is_element_present,
@@ -10,15 +11,15 @@ from pytest_bdd import (
     given,
     scenario,
     then,
-    when,
+    when
 )
 import pytest
 pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1346.feature', 'Apps Page - Stop an App')
-def test_apps_page__stop_an_app():
-    """Apps Page - Stop an App."""
+@scenario('features/NAS-T1346.feature', 'Apps Page - Validate machinaris')
+def test_apps_page__validate_machinaris():
+    """Apps Page - Validate machinaris."""
 
 
 @given('the browser is open, navigate to the SCALE URL, and login')
@@ -48,20 +49,15 @@ def on_the_dashboard_click_on_apps(driver):
     driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
 
 
-@then('the Apps page load, open installed applications')
-def the_apps_page_load_open_installed_applications(driver):
-    """the Apps page load, open installed applications."""
+@then('Stop Chia from running')
+def stop_chia_from_running(driver):
+    """Stop Chia from running."""
     if is_element_present(driver, '//mat-ink-bar[@style="visibility: visible; left: 0px; width: 183px;"]') is False:
         assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
         driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
         assert wait_on_element(driver, 7, '//h3[contains(.,"No Applications Installed")]')
-
-
-@then('click the stop button and confirm')
-def click_the_stop_button_and_confirm(driver):
-    """click the stop button and confirm."""
-    assert wait_on_element(driver, 20, '//mat-card[contains(.,"collabora")]//span[contains(.,"Stop")]', 'clickable')
-    driver.find_element_by_xpath('//mat-card[contains(.,"collabora")]//span[contains(.,"Stop")]').click()
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"chia")]//span[contains(.,"Stop")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"chia")]//span[contains(.,"Stop")]').click()
 
 
 @then('Verify the application has stopped')
@@ -69,4 +65,122 @@ def verify_the_application_has_stopped(driver):
     """Verify the application has stopped."""
     assert wait_on_element(driver, 5, '//h1[contains(.,"Stopping")]')
     assert wait_on_element_disappear(driver, 60, '//h1[contains(.,"Stopping")]')
-    assert wait_on_element(driver, 15, '//mat-card[contains(.,"collabora-test")]//span[contains(.,"STOPPED ")]')
+    assert wait_on_element(driver, 15, '//mat-card[contains(.,"chia-test")]//span[contains(.,"STOPPED ")]')
+
+
+@then('open available applications')
+def open_available_applications(driver):
+    """open available applications."""
+    assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
+    driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
+    assert wait_on_element(driver, 7, '//div[contains(.,"Available Applications")]')
+
+
+@then('click install')
+def click_install(driver):
+    """click install."""
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    assert wait_on_element(driver, 20, '//mat-card[contains(.,"machinaris")]//span[contains(.,"Install")]', 'clickable')
+    driver.find_element_by_xpath('//mat-card[contains(.,"machinaris")]//span[contains(.,"Install")]').click()
+    assert wait_on_element(driver, 5, '//*[contains(.,"Please wait")]')
+    assert wait_on_element_disappear(driver, 30, '//*[contains(.,"Please wait")]')
+
+
+@then('set application name')
+def set_application_name(driver):
+    """set application name."""
+    assert wait_on_element(driver, 7, '//h3[contains(.,"machinaris")]')
+    assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
+    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('machinaris-test')
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
+
+
+@then('set networking')
+def set_networking(driver):
+    """set networking."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Networking"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Networking"]').click()
+
+
+@then('set machinaris configuration')
+def set_machinaris_configuration(driver):
+    """set machinaris configuration."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Machinaris Configuration"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Machinaris Configuration"]').click()
+
+
+@then('set storage')
+def set_storage(driver):
+    """set storage."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Storage"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
+
+
+@then('set Machinaris Environment Variables')
+def set_machinaris_environment_variables(driver):
+    """set Machinaris Environment Variables."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Machinaris Environment Variables"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Machinaris Environment Variables"]').click()
+
+
+@then('set Resource Limits')
+def set_resource_limits(driver):
+    """set Resource Limits."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Resource Limits"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Resource Limits"]').click()
+
+
+@then('set Configure Coins')
+def set_configure_coins(driver):
+    """set Configure Coins."""
+    assert wait_on_element(driver, 10, '//mat-checkbox[@ix-auto="checkbox__Enable Flax"]', 'clickable')
+    driver.find_element_by_xpath('//mat-checkbox[@ix-auto="checkbox__Enable Flax"]').click()
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Configure Coins"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Configure Coins"]').click()
+
+
+@then('confirm options')
+def confirm_options(driver):
+    """confirm options."""
+    assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
+
+    assert wait_on_element(driver, 5, '//*[contains(.,"Installing")]')
+    assert wait_on_element_disappear(driver, 45, '//*[contains(.,"Installing")]')
+
+
+@then('confirm installation is successful')
+def confirm_installation_is_successful(driver):
+    """confirm installation is successful."""
+    assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+    driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
+    if is_element_present(driver, '//mat-card[contains(.,"machinaris-test")]//span[@class="status active"]') is False:
+        assert wait_on_element(driver, 20, '//strong[contains(.,"machinaris-test")]')
+        assert wait_on_element(driver, 20, '//strong[contains(.,"machinaris-test")]', 'clickable')
+        driver.find_element_by_xpath('//strong[contains(.,"machinaris-test")]').click()
+        if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
+            assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"machinaris-test")]')
+        assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
+        driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
+        while is_element_present(driver, '//div[(normalize-space(text())="Created container machinaris")]') is False:
+            time.sleep(2)
+            assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
+            # make sure Please wait pop up is gone before continuing.
+            if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
+                assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+        else:
+            assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
+            driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
+            time.sleep(45)  # Because of slow start up times, this takes another 10-20 second to switch from "Deploying to Active"  So we can either flip the page constantly or just wait and give it time.
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
+            assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
+            driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
+            assert wait_on_element(driver, 600, '//mat-card[contains(.,"machinaris-test")]//span[@class="status active"]')
+    else:
+        assert wait_on_element(driver, 600, '//mat-card[contains(.,"machinaris-test")]//span[@class="status active"]')
