+++ tests/bdd/scale/test_NAS_T1354.py
@@ -5,7 +5,7 @@ import time
 from function import (
     wait_on_element,
     is_element_present,
-    wait_on_element_disappear,
+    wait_on_element_disappear
 )
 from pytest_bdd import (
     given,
@@ -13,13 +13,11 @@ from pytest_bdd import (
     then,
     when
 )
-import pytest
-pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1354.feature', 'Apps Page - Validate ipfs')
-def test_apps_page__validate_ipfs():
-    """Apps Page - Validate ipfs."""
+@scenario('features/NAS-T1354.feature', 'Apps Page - Validate removing a Catalog')
+def test_apps_page__validate_removing_a_catalog():
+    """Apps Page - Validate removing a Catalog."""
 
 
 @given('the browser is open, navigate to the SCALE URL, and login')
@@ -49,105 +47,37 @@ def on_the_dashboard_click_on_apps(driver):
     driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
 
 
-@then('the Apps page load, open available applications')
-def the_apps_page_load_open_available_applications(driver):
-    """the Apps page load, open available applications."""
-    assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
-    driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
-    assert wait_on_element(driver, 7, '//div[contains(.,"Available Applications")]')
+@then('when the Apps page loads, open manage catalogs')
+def when_the_apps_page_loads_open_manage_catalogs(driver):
+    """when the Apps page loads, open manage catalogs."""
+    assert wait_on_element(driver, 10, '//div[contains(text(),"Manage Catalogs")]', 'clickable')
+    driver.find_element_by_xpath('//div[contains(text(),"Manage Catalogs")]').click()
+    assert wait_on_element(driver, 7, '//div[contains(.,"Manage Catalogs")]')
 
 
-@then('click install')
-def click_install(driver):
-    """click install."""
-    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    assert wait_on_element(driver, 20, '//mat-card[contains(.,"ipfs")]//span[contains(.,"Install")]', 'clickable')
-    driver.find_element_by_xpath('//mat-card[contains(.,"ipfs")]//span[contains(.,"Install")]').click()
-    if is_element_present(driver, '//*[contains(.,"Please wait")]'):
-        assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
+@then('click three dots icon for Truecharts and select delete')
+def click_three_dots_icon_for_truecharts_and_select_delete(driver):
+    """click three dots icon for Truecharts and select delete."""
+    time.sleep(5)  # we have to wait for the page to settle down and the card to fully load
+    assert wait_on_element(driver, 10, '//tr[contains(.,"CUSTOMCHART")]//mat-icon[contains(.," more_vert ")]', 'clickable')
+    driver.find_element_by_xpath('//tr[contains(.,"CUSTOMCHART")]//mat-icon[contains(.," more_vert ")]').click()
+    assert wait_on_element(driver, 10, '//button[@ix-auto="action__CUSTOMCHART_Delete"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="action__CUSTOMCHART_Delete"]').click()
 
 
-@then('set application name')
-def set_application_name(driver):
-    """set application name."""
-    assert wait_on_element(driver, 7, '//h3[contains(.,"ipfs")]')
-    assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
-    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
-    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('ipfs-test')
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
+@then('confirm the confirmation')
+def confirm_the_confirmation(driver):
+    """confirm the confirmation."""
+    assert wait_on_element(driver, 2, '//mat-checkbox[@ix-auto="checkbox__CONFIRM"]', 'clickable')
+    driver.find_element_by_xpath('//mat-checkbox[@ix-auto="checkbox__CONFIRM"]').click()
+    assert wait_on_element(driver, 10, '//button[@ix-auto="button__DELETE"]', 'clickable')
+    driver.find_element_by_xpath('//button[@ix-auto="button__DELETE"]').click()
+    time.sleep(0.5)
+    assert wait_on_element_disappear(driver, 25, '//h6[contains(.,"Please wait")]')
 
 
-@then('set workload configuration')
-def set_workload_configuration(driver):
-    """set workload configuration."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Workload Configuration"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Workload Configuration"]').click()
-
-
-@then('set storage')
-def set_storage(driver):
-    """set storage."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Storage"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
-
-
-@then('IPFS Configuration')
-def ipfs_configuration(driver):
-    """IPFS Configuration."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_IPFS Configuration"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_IPFS Configuration"]').click()
-
-
-@then('Advanced DNS Settings')
-def advanced_dns_settings(driver):
-    """Advanced DNS Settings."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Advanced DNS Settings"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Advanced DNS Settings"]').click()
-
-
-@then('confirm options')
-def confirm_options(driver):
-    """confirm options."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
-
-    assert wait_on_element(driver, 5, '//*[contains(.,"Installing")]')
-    assert wait_on_element_disappear(driver, 45, '//*[contains(.,"Installing")]')
-
-
-@then('confirm installation is successful')
-def confirm_installation_is_successful(driver):
-    """confirm installation is successful."""
-    assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
-    driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    if is_element_present(driver, '//mat-card[contains(.,"ipfs-test")]//span[@class="status active"]') is False:
-        assert wait_on_element(driver, 20, '//strong[contains(.,"ipfs-test")]')
-        assert wait_on_element(driver, 20, '//strong[contains(.,"ipfs-test")]', 'clickable')
-        driver.find_element_by_xpath('//strong[contains(.,"ipfs-test")]').click()
-        assert wait_on_element(driver, 5, '//*[contains(.,"Please wait")]')
-        if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
-            assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
-        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"ipfs-test")]')
-        assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
-        driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
-        while is_element_present(driver, '//div[(normalize-space(text())="Started container ipfs")]') is False:
-            time.sleep(2)
-            assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
-            driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
-            # make sure Please wait pop up is gone before continuing.
-            if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
-                assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
-        else:
-            assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
-            driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
-            time.sleep(1)  # wait for popup to close
-            # we have to change tab for UI to refresh
-            assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
-            driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
-            assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
-            driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-            assert wait_on_element(driver, 500, '//mat-card[contains(.,"ipfs-test")]//span[@class="status active"]')
-    else:
-        assert wait_on_element(driver, 500, '//mat-card[contains(.,"ipfs-test")]//span[@class="status active"]')
+@then('confirm deletion is successful')
+def confirm_deletion_is_successful(driver):
+    """confirm deletion is successful."""
+    assert wait_on_element_disappear(driver, 10, '//div[text()="CUSTOMCHART"]')
+    assert is_element_present(driver, '//div[text()="CUSTOMCHART"]') is False
