+++ tests/bdd/scale/test_NAS_T1356.py
@@ -18,7 +18,7 @@ import pytest
 pytestmark = [pytest.mark.debug_test]
 
 
-@scenario('features/NAS-T1539.feature', 'Apps Page - Validate adding TrueCommand as a custom app')
+@scenario('features/NAS-T1356.feature', 'Apps Page - Validate adding TrueCommand as a custom app')
 def test_apps_page__validate_adding_truecommand_as_a_custom_app():
     """Apps Page - Validate adding TrueCommand as a custom app."""
