--- tests/bdd/scale/test_NAS_T1532.py
@@ -1,160 +1,0 @@
-# coding=utf-8
-"""SCALE UI: feature tests."""
-
-import time
-from function import (
-    wait_on_element,
-    is_element_present,
-    wait_on_element_disappear,
-)
-from pytest_bdd import (
-    given,
-    scenario,
-    then,
-    when,
-)
-import pytest
-pytestmark = [pytest.mark.debug_test]
-
-
-@scenario('features/NAS-T1532.feature', 'Apps Page - Validate nextcloud')
-def test__apps_page__validate_nextcloud():
-    """SCALE UI: Apps Page - Validate nextcloud."""
-
-
-@given('the browser is open, navigate to the SCALE URL, and login')
-def the_browser_is_open_navigate_to_the_scale_url_and_login(driver, nas_ip, root_password):
-    """the browser is open, navigate to the SCALE URL, and login."""
-    if nas_ip not in driver.current_url:
-        driver.get(f"http://{nas_ip}")
-        assert wait_on_element(driver, 10, '//input[@data-placeholder="Username"]')
-    if not is_element_present(driver, '//mat-list-item[@ix-auto="option__Dashboard"]'):
-        assert wait_on_element(driver, 10, '//input[@data-placeholder="Username"]')
-        driver.find_element_by_xpath('//input[@data-placeholder="Username"]').clear()
-        driver.find_element_by_xpath('//input[@data-placeholder="Username"]').send_keys('root')
-        driver.find_element_by_xpath('//input[@data-placeholder="Password"]').clear()
-        driver.find_element_by_xpath('//input[@data-placeholder="Password"]').send_keys(root_password)
-        assert wait_on_element(driver, 5, '//button[@name="signin_button"]')
-        driver.find_element_by_xpath('//button[@name="signin_button"]').click()
-    else:
-        assert wait_on_element(driver, 5, '//mat-list-item[@ix-auto="option__Dashboard"]', 'clickable')
-        driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Dashboard"]').click()
-
-
-@when('on the Dashboard, click on apps')
-def on_the_dashboard_click_on_apps(driver):
-    """on the Dashboard, click on apps."""
-    assert wait_on_element(driver, 10, '//span[contains(.,"Dashboard")]')
-    assert wait_on_element(driver, 10, '//mat-list-item[@ix-auto="option__Apps"]', 'clickable')
-    driver.find_element_by_xpath('//mat-list-item[@ix-auto="option__Apps"]').click()
-
-
-@then('the Apps page load, open available applications')
-def the_apps_page_load_open_available_applications(driver):
-    """the Apps page load, open available applications."""
-    assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
-    driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
-    assert wait_on_element(driver, 7, '//div[contains(.,"Available Applications")]')
-
-
-@then('click install')
-def click_install(driver):
-    """click install."""
-    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    assert wait_on_element(driver, 20, '//mat-card[contains(.,"nextcloud")]//span[contains(.,"Install")]', 'clickable')
-    driver.find_element_by_xpath('//mat-card[contains(.,"nextcloud")]//span[contains(.,"Install")]').click()
-    if is_element_present(driver, '//*[contains(.,"Please wait")]'):
-        assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
-
-
-@then('set application name')
-def set_application_name(driver):
-    """set application name."""
-    assert wait_on_element(driver, 7, '//h3[contains(.,"nextcloud")]')
-    assert wait_on_element(driver, 7, '//input[@ix-auto="input__Application Name"]')
-    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').clear()
-    driver.find_element_by_xpath('//input[@ix-auto="input__Application Name"]').send_keys('nextcloud-test')
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Application Name"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Application Name"]').click()
-
-
-@then('set nexcloud configuration')
-def set_nexcloud_configuration(driver):
-    """set nexcloud configuration."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Nextcloud Configuration"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Nextcloud Configuration"]').click()
-
-
-@then('set storage')
-def set_storage(driver):
-    """set storage."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Storage"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Storage"]').click()
-
-
-@then('on CronJob configuration click next')
-def on_cronjob_configuration_click_next(driver):
-    """on CronJob configuration click next."""
-    wait_on_element(driver, 5, '//mat-step-header[contains(.,"CronJob configuration") and @tabindex="0"]')
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_CronJob configuration"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_CronJob configuration"]').click()
-
-
-@then('set Scaling Upgrade Policy')
-def set_scaling_upgrade_policy(driver):
-    """set Scaling Upgrade Policy."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Scaling/Upgrade Policy"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Scaling/Upgrade Policy"]').click()
-
-
-@then('Advanced DNS Settings')
-def advanced_dns_settings(driver):
-    """Advanced DNS Settings."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__NEXT_Advanced DNS Settings"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__NEXT_Advanced DNS Settings"]').click()
-
-
-@then('confirm options')
-def confirm_options(driver):
-    """confirm options."""
-    assert wait_on_element(driver, 7, '//button[@ix-auto="button__SAVE"]', 'clickable')
-    driver.find_element_by_xpath('//button[@ix-auto="button__SAVE"]').click()
-
-    assert wait_on_element(driver, 10, '//*[contains(.,"Installing")]')
-    assert wait_on_element_disappear(driver, 60, '//*[contains(.,"Installing")]')
-
-
-@then('confirm installation is successful')
-def confirm_installation_is_successful(driver):
-    """confirm installation is successful."""
-    assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
-    driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-    time.sleep(2)  # we have to wait for the page to settle down and the card to fully load
-    if is_element_present(driver, '//mat-card[contains(.,"nextcloud-test")]//span[@class="status active"]') is False:
-        assert wait_on_element(driver, 20, '//strong[contains(.,"nextcloud-test")]')
-        assert wait_on_element(driver, 20, '//strong[contains(.,"nextcloud-test")]', 'clickable')
-        driver.find_element_by_xpath('//strong[contains(.,"nextcloud-test")]').click()
-        if wait_on_element(driver, 5, '//*[contains(.,"Please wait")]'):
-            assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
-        assert wait_on_element(driver, 10, '//div[@class="logo-container" and contains(.,"nextcloud-test")]')
-        assert wait_on_element(driver, 10, '//mat-panel-title[contains(.,"Application Events")]', 'clickable')
-        driver.find_element_by_xpath('//mat-panel-title[contains(.,"Application Events")]').click()
-        while is_element_present(driver, '//div[(normalize-space(text())="Started container nextcloud")]') is False:
-            time.sleep(2)
-            assert wait_on_element(driver, 10, '//span[contains(.,"Refresh Events")]', 'clickable')
-            driver.find_element_by_xpath('//span[contains(.,"Refresh Events")]').click()
-            # make sure Please wait pop up is gone before continuing.
-            if wait_on_element(driver, 3, '//*[contains(.,"Please wait")]'):
-                assert wait_on_element_disappear(driver, 10, '//*[contains(.,"Please wait")]')
-        else:
-            assert wait_on_element(driver, 10, '//span[contains(.,"Close")]', 'clickable')
-            driver.find_element_by_xpath('//span[contains(.,"Close")]').click()
-            time.sleep(1)  # wait for popup to close
-            # we have to change tab for UI to refresh
-            assert wait_on_element(driver, 10, '//div[contains(text(),"Available Applications")]', 'clickable')
-            driver.find_element_by_xpath('//div[contains(text(),"Available Applications")]').click()
-            assert wait_on_element(driver, 10, '//div[contains(text(),"Installed Applications")]', 'clickable')
-            driver.find_element_by_xpath('//div[contains(text(),"Installed Applications")]').click()
-            assert wait_on_element(driver, 500, '//mat-card[contains(.,"nextcloud-test")]//span[@class="status active"]')
-    else:
-        assert wait_on_element(driver, 500, '//mat-card[contains(.,"nextcloud-test")]//span[@class="status active"]')
